Restructed API calls in preparation for additional local API functions.
