+++ packages/esm-patient-list-app/src/patientListData/index.ts
@@ -1,3 +1,4 @@
+export * from './api-remote';
+export * from './api-local';
 export * from './api';
-export * from './localPatientLists';
-export * from './hooks';
+export * from './queries';
